Challenge/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
passlib[bcrypt]
alembic
python-jose[cryptography]
pyjwt
decouple
